V90CP: the write cursor is unsigned, and bitsToInfo is what says so
+0xcac was `int` because nothing read so far forced either reading -- every
use in the class is a plain 32-bit load or store, which is the same
instruction signed or unsigned.  `bitsToInfo` forces it twice: the bit-vector
bound is `cmp $0x2edf` / `ja`, an unsigned above where a signed `int`
comparison is `jg`, and `word_cac % 6` is the 0xaaaaaaab reciprocal with a
plain `shr`, which is the unsigned magic and not the signed one.

Measured before it was believed, because a retype is not free the way a rename
is: header-only, no new code, `make phase` 208 passed / 0 failed, `compare.py`
unmoved at 363 identical / 72 same-size / 1017 compared / 82.1%.

Finding 4362.

+++ src/pump/v90/V90CP.cpp
@@ -761,7 +761,7 @@ V90CP::infoToBits()
 
 	start = (unsigned int)word_cac;
 	bits[start] = 0;
-	word_cac = (int)(start + 1);
+	word_cac = start + 1;
 	nbits = start + 0x11;
 	word_3bb0 = nbits;
 
@@ -798,7 +798,7 @@ V90CP::infoToBits()
 		pos++;
 	}
 	bits[pos] = 0;
-	word_cac = (int)(pos + 1);
+	word_cac = pos + 1;
 
 	/* calcSequenceLength, inlined: round the bit count up to a whole
 	 * number of +0x3ba8 and record it at +0x3bac. */
